backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
